requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
alembic==1.8.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
banal==1.0.6
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
datasets==2.6.1
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
efficientnet-pytorch==0.7.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.0
fonttools==4.37.4
frozenlist==1.3.1
fsspec==2022.10.0
greenlet==1.1.3.post0
huggingface-hub==0.10.1
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipykernel==6.17.0
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
jupyter==1.0.0
jupyter-client==7.4.4
jupyter-console==6.4.4
jupyter-core==4.11.2
jupyter-server==1.21.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
Mako==1.2.3
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.2
multiprocess==0.70.13
nbclassic==0.4.7
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
nltk==3.7
notebook==6.5.2
notebook-shim==0.2.2
numpy==1.23.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil-resolve-name==1.3.10
prometheus-client==0.15.0
prompt-toolkit==3.0.31
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.2.1
regex==2022.10.31
requests==2.28.1
responses==0.18.0
scikit-learn==1.1.2
scipy==1.9.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.42
stack-data==0.6.0
terminado==0.17.0
threadpoolctl==3.1.0
timm==0.6.11
tinycss2==1.2.1
tokenizers==0.13.2
torch==1.12.1
torchinfo==1.7.1
torchmetrics==0.10.2
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.24.0
typing-extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
widgetsnbextension==4.0.3
xxhash==3.1.0
yapf==0.32.0
yarl==1.8.1
zipp==3.9.0
